+++ dart/test/event_processor/enricher/io_enricher_test.dart
@@ -18,9 +18,9 @@ void main() {
     fixture = Fixture();
   });
 
-  test('adds dart runtime', () {
+  test('adds dart runtime', () async {
     final enricher = fixture.getSut();
-    final event = enricher.apply(SentryEvent(), Hint());
+    final event = await enricher.apply(SentryEvent(), Hint());
 
     expect(event?.contexts.runtimes, isNotEmpty);
     final dartRuntime = event?.contexts.runtimes
@@ -31,12 +31,12 @@ void main() {
     expect(Platform.version, contains(dartRuntime.version.toString()));
   });
 
-  test('does add to existing runtimes', () {
+  test('does add to existing runtimes', () async {
     final runtime = SentryRuntime(name: 'foo', version: 'bar');
     var event = SentryEvent(contexts: Contexts(runtimes: [runtime]));
     final enricher = fixture.getSut();
 
-    event = enricher.apply(event, Hint())!;
+    event = (await enricher.apply(event, Hint()))!;
 
     expect(event.contexts.runtimes.contains(runtime), true);
     // second runtime is Dart runtime
@@ -45,10 +45,10 @@ void main() {
 
   group('adds device, os and culture', () {
     for (final hasNativeIntegration in [true, false]) {
-      test('native=$hasNativeIntegration', () {
+      test('native=$hasNativeIntegration', () async {
         final enricher =
             fixture.getSut(hasNativeIntegration: hasNativeIntegration);
-        final event = enricher.apply(SentryEvent(), Hint());
+        final event = await enricher.apply(SentryEvent(), Hint());
 
         expect(event?.contexts.device, isNotNull);
         expect(event?.contexts.operatingSystem, isNotNull);
@@ -57,31 +57,31 @@ void main() {
     }
   });
 
-  test('device has no name if sendDefaultPii = false', () {
+  test('device has no name if sendDefaultPii = false', () async {
     final enricher = fixture.getSut();
-    final event = enricher.apply(SentryEvent(), Hint());
+    final event = await enricher.apply(SentryEvent(), Hint());
 
     expect(event?.contexts.device?.name, isNull);
   });
 
-  test('device has name if sendDefaultPii = true', () {
+  test('device has name if sendDefaultPii = true', () async {
     final enricher = fixture.getSut(includePii: true);
-    final event = enricher.apply(SentryEvent(), Hint());
+    final event = await enricher.apply(SentryEvent(), Hint());
 
     expect(event?.contexts.device?.name, isNotNull);
   });
 
-  test('culture has locale and timezone', () {
+  test('culture has locale and timezone', () async {
     final enricher = fixture.getSut();
-    final event = enricher.apply(SentryEvent(), Hint());
+    final event = await enricher.apply(SentryEvent(), Hint());
 
     expect(event?.contexts.culture?.locale, isNotNull);
     expect(event?.contexts.culture?.timezone, isNotNull);
   });
 
-  test('os has name and version', () {
+  test('os has name and version', () async {
     final enricher = fixture.getSut();
-    final event = enricher.apply(SentryEvent(), Hint());
+    final event = await enricher.apply(SentryEvent(), Hint());
 
     expect(event?.contexts.operatingSystem?.name, isNotNull);
     if (Platform.isLinux) {
@@ -158,9 +158,9 @@ void main() {
     });
   });
 
-  test('adds Dart context with PII', () {
+  test('adds Dart context with PII', () async {
     final enricher = fixture.getSut(includePii: true);
-    final event = enricher.apply(SentryEvent(), Hint());
+    final event = await enricher.apply(SentryEvent(), Hint());
 
     final dartContext = event?.contexts['dart_context'];
     expect(dartContext, isNotNull);
@@ -171,9 +171,9 @@ void main() {
     // package_config and executable_arguments are optional
   });
 
-  test('adds Dart context without PII', () {
+  test('adds Dart context without PII', () async {
     final enricher = fixture.getSut(includePii: false);
-    final event = enricher.apply(SentryEvent(), Hint());
+    final event = await enricher.apply(SentryEvent(), Hint());
 
     final dartContext = event?.contexts['dart_context'];
     expect(dartContext, isNotNull);
@@ -185,7 +185,7 @@ void main() {
     // and Platform is not mockable
   });
 
-  test('does not override event', () {
+  test('does not override event', () async {
     final fakeEvent = SentryEvent(
       contexts: Contexts(
         device: SentryDevice(
@@ -207,7 +207,7 @@ void main() {
       hasNativeIntegration: false,
     );
 
-    final event = enricher.apply(fakeEvent, Hint());
+    final event = await enricher.apply(fakeEvent, Hint());
 
     // contexts.device
     expect(